Using a plugin for vite to work with istanbul
