+++ babel.config.json
@@ -9,9 +9,5 @@
       }
     ]
   ],
-  "env": {
-    "test": {
-      "plugins": ["istanbul"]
-    }
-  }
+  "plugins": ["istanbul"]
 }
